Remove duplicity using an abstract class for AwsKmsTestKekManager
